requirements.txt
pandas==2.1.2
requests==2.31.0
PyYAML==6.0.1
scipy==1.11.3
autogluon.tabular==0.8.2
scikit-learn==1.2.2